+++ src/packerlicious/community/__init__.py
@@ -1,0 +1,3 @@
+"""
+packer plugins that are not distributed with packer core.
+"""
